% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get}
\alias{cache_get}
\title{Retrieve from cache}
\usage{
cache_get(function_name, function_call, cache)
}
\arguments{
\item{function_name}{function to retrieve against}

\item{function_call}{function call to retrieve against}

\item{cache}{a Cache object}
}
\value{
the result of evaluating \code{function_call} (either directly or
from the cache)
}
\description{
Retrieve from cache
}

