% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set}
\alias{cache_set}
\title{Set an entry in the cache}
\usage{
cache_set(function_name, arg_hash, result, cache)
}
\arguments{
\item{function_name}{function_name to set against}

\item{arg_hash}{the hash of the arguments}

\item{result}{the result to cache}

\item{cache}{the Cache object to cache it in}
}
\description{
Set an entry in the cache
}

