% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached_call}
\alias{cached_call}
\title{Return the output of a function using a cache}
\usage{
cached_call(fun, ..., cache = NULL)
}
\arguments{
\item{fun}{function to call}

\item{...}{arguments to \code{fun} (excluding cache). Both named and positional
arguments are permitted.}

\item{cache}{a Cache object}
}
\value{
The output of \code{fun}
}
\description{
Check \code{cache} for the cached output of \code{fun} for the arguments
\code{...}. If a cache entry is found, return it. Otherwise call \code{fun}
with arguments \code{...} (and \code{cache} if cache is a parameter of
\code{fun}), cache the result and then return it.
}
\examples{
cache <- create_cache()

result <- cached_call(mean, x=1:10, cache=cache)
result2 <- cached_call(mean, 1:10, cache=cache)

}

