% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheMan.R
\docType{package}
\name{cacheMan-package}
\alias{cacheMan}
\alias{cacheMan-package}
\title{An in-memory cache manager for R}
\description{
cacheMan is a simple in-memory cache manager for R. Functions can be wrapped
to allow them to retrieve results from memory rather than recalculating them
when their parameters have not changed.
}

