% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheUtils.R
\name{hash_object}
\alias{hash_object}
\title{Hash an object}
\usage{
hash_object(object, cache)
}
\arguments{
\item{object}{an object to hash}

\item{cache}{the Cache object to take the hash function from}
}
\value{
a hash of \code{object}
}
\description{
Hash an object
}

