% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheUtils.R
\name{print.Cache}
\alias{print.Cache}
\title{Print a cache function}
\usage{
\method{print}{Cache}(cache)
}
\arguments{
\item{cache}{a Cache object}
}
\description{
Print a cache function
}
\examples{
cache <- create_cache()
print(cache)
}

