% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{create_cache}
\alias{create_cache}
\title{Create an empty cache}
\usage{
create_cache(function_cache_size = 1, hash_algo = "crc32")
}
\arguments{
\item{function_cache_size}{number of results to store for each cached
function. When this limit is reached, new entries in the cache will
overwrite the oldest entry for the relevant function.}

\item{hash_algo}{the hash algorithm to use.}
}
\value{
a Cache object
}
\description{
Create an empty cache
}
\examples{
cache <- create_cache()
}

